requirements.txt
streamlit==1.27.0
openai==0.27.8
python-dotenv==1.0.0
pydantic==1.10.8
